MVC pattern for Search Screen
